•Доработан интерфейс; •Добавлен enum для советов показателя ИМТ; •Добавлен класс с новым окном для показания пользователю совета по уровню ИМТ; •Исправлены графические и синтаксические мелкие баги.

+++ src/main/java/org/view/Processing.java
@@ -1,7 +1,7 @@
 package org.view;
 
 public class Processing {
-    static public float calculateBMI(float height, float weight){
+    static public double calculateBMI(double height, double weight){
         return weight / (height * height);
     }
 }
